Fix missing/broken $(CONFIG_UPDATE) calls
This patch fixes a broken use of $(CONFIG_UPDATE) in bash.mk
(config.sub and config.guess is stored in the "support" subdirectory,
not directly under $(BASH_DIR)) and adds $(CONFIG_UPDATE) to
several other makefiles.

There are probably other cases that need to be fixed up as well,
but these are the ones that broke for me when compiling for AVR32.

Signed-off-by: Haavard Skinnemoen <hskinnemoen@atmel.com>