Add ATNGW100 target device board support
This patch adds support for the ATNGW100 reference design board from Atmel. The
board is based on the AVR32 AT32AP7000 device, 8 MB NOR flash, 8 MB DataFlash
and 32 MB 16-bit SDRAM.

Included in the patch is default configurations for Busybox, Linux and
Buildroot. There is also a target_skeleton for the basic files needed to boot
the system with the default Buildroot configuration.

Tested on a ATNGW100 revision C.

Signed-off-by: Hans-Christian Egtvedt <hcegtvedt@atmel.com>