qtopia4: add license choice to Qtopia 4 Core
This patch adds a choice of license type, comercial or GPL, to Kconfig. It will
also give you the possibility to approve the GPL license on Qtopia 4 Core so
you do not have to ask yes twice during compilation.

This will greatly reduce the need to attend the compilation, and necessary for
automatic builds.

Signed-off-by: Hans-Christian Egtvedt <hcegtvedt@atmel.com>