u-boot: add U-Boot bootloader
This patch adds Das U-Boot to Buildroot. It will add a general target
bootloader directory with U-Boot in a sub directory from there.

When using U-Boot it is also assumed you are making a kernel image with the
U-Boot tool mkimage, this is built automatically when choosing U-Boot and
install into the staging_dir as part of the toolchain.

U-Boot is built and configured from menuconfig, where you have to specify the
target board config which is used for make.

Signed-off-by: Hans-Christian Egtvedt <hcegtvedt@atmel.com>