Add libaio package
Make it possible to build the libaio library for asynchronous I/O.

Since the upstream package seems to be unmaintained, I grabbed the
latest Debian source package, imported it into a git repository and
ported it to AVR32. The resulting tree was uploaded to avr32linux.org,
which is where buildroot grabs it from.

Signed-off-by: Haavard Skinnemoen <hskinnemoen@atmel.com>