Add gadgetfs-test package
This makes it possible to build the gadgetfs test program from
http://www.linux-usb.org/gadget/. Since the program is distributed as
discrete files without a Makefile or anything, a tarball with all the
source files and a simple Makefile has been uploaded to
avr32linux.org. This is where buildroot grabs it from.

Currently, the headers exported by the Linux kernel doesn't include
usb_gadgetfs.h. A patch that fixes this (and moves the header in the
process) is in -mm, but we need to backport it to our own kernel
before this program will work.

Signed-off-by: Haavard Skinnemoen <hskinnemoen@atmel.com>
Signed-off-by: Hans-Christian Egtvedt <hcegtvedt@atmel.com>