qtopia4: Add more image support options, cleanup makefile and comments
This patch adds options for the image support available in Qtopia, more
specific GIF and MNG. It also will depend on libpng if the user has selected to
compile that library.

This patch also tries to clean up the configure and installing of Qtopia, both
the library for staging_dir and the library for the target. I have tested this
on AVR32 and it compiles successfull and I can also compile other Qt
applications using this setup.

A test on other platforms would have been nice.

Signed-off-by: Hans-Christian Egtvedt <hcegtvedt@atmel.com>