AVR32 boards update startup scripts, devices and makefiles
This patch updates the init.d files for NGW100 to not output anything to
/dev/null.

The Makefiles for both STK1000 and NGW100 will now not interfer with each
others.

Device tables is updated to create the /dev/null node for the boot process.

Updated default Buildroot configuration is also added for both boards.

Signed-off-by: Hans-Christian Egtvedt <hcegtvedt@atmel.com>