u-boot: added configuration of target header file
This patch adds support for changing network setup, bootarguments and bootcmd
before u-boot is built.

It also adds support for a user specified header file configuration which is
copied from a given location into the target board config header file.

This allows the user to configure U-Boot as he prefers.

Signed-off-by: Hans-Christian Egtvedt <hcegtvedt@atmel.com>