dependencies: check if the compiler is executable before assuming it works
This patch adds a simple check in the dependencies shell script, checking if
the compiler (C and C++) is actually executable before assuming they exist and
try to run with --version.

On a Ubuntu system the script failed to detect that a compiler was not
installed, and inform the user correctly.

Signed-off-by: Hans-Christian Egtvedt <hcegtvedt@atmel.com>