Rename Documentation stuff packages to AVR32 stuff and rearrange the menu
This patch renames the "Documentation stuff" entry under packages in menuconfig
to "AVR32 stuff". Today it only content is the Wiki documentation.

The menu entry is also moved to the top of the list and will only be visible if
AVR32 architecture is chosen. This should make it easier for AVR32 users to
spot and disable if they do not want the Wiki on their target device.

Signed-off-by: Hans-Christian Egtvedt <hcegtvedt@atmel.com>