jffs2root: add support for multiple partitions
This patch adds a possibility to split the JFFS2 images into multiple
partitions. This is very useful if you have more than one flash device you want
to populate.

For example could you have a paralell flash for the / file system while use a
serial flash for /usr.

The patch contains an example setup file for JFFS2 partitions. The default
place to look for the partition setup file is in
BR2_BOARD_PATH/jffs2_partitions.txt.

Signed-off-by: Hans-Christian Egtvedt <hcegtvedt@atmel.com>