Remove samba, mpg123 and madplay from the default config for ATSTK1002
Samba is hard to squeeze in on the BSP for ATSTK1002 because it only has 8 MB
of RAM in total, usually leaving roughly 6 MB for the running system after the
kernel and framebuffer is reserved.

mpg123 and madplay is not needed as long as MPlayer is present. MPlayer is also
more flexible with audio out, which is handy when using the AT73C213 as sound
sink due to the fixed bitrate this device needs.

Signed-off-by: Hans-Christian Egtvedt <hcegtvedt@atmel.com>