Add oprofile package
This adds an oprofile package to buildroot, along with AVR32 support
and two quick fixes.

This should be considered experimental at this point, since
  * avr32 oprofile support in the kernel hasn't been merged yet
  * gcc 4.2.1 miscompiles libbfd, causing opreport to crash
  * the autoconf patch is just a quick hack
  * enabling oprofile silently compiles and installs binutils for the
    target.
  * target-binutils doesn't install the necessary headers and libraries
    into staging_dir, so additional kludges are needed.

But with a bit of hacking around, it actually works. A 25 second
profiling session with iperf showed that 25% of the time is spent in
memcpy(), which makes sense, and it confirms that my (also
experimental) patchset that eliminates memcpy from the macb receive
path is on the right track.

Signed-off-by: Haavard Skinnemoen <hskinnemoen@atmel.com>