Remove U-Boot MAC and update bootcmd for AVR32 boards
This patch removes a given MAC for AVR32 boards since this should be a proper
value set by the user. It is also possible to set this MAC after programming
the bootloader.

The bootcmd is also updated to load the kernel at address 0x90300000 instead of
0x90250000, since the latter address may be to small for some large kernels
(i.e. bigger than 2.4 MB extracted).

Signed-off-by: Hans-Christian Egtvedt <hcegtvedt@atmel.com>