linux: strip variables before checking if they are empty and add u-boot dependency
This patch strips the LINUX26_BINLOC and LINUX26_FORMAT before checking their
value, this ensures that they will work as expected.

The patch will also add u-boot dependency to LINUX26_TARGETS if u-boot
bootloader is selected in kconfig.

The TARGET_PATH is also added when building the kernel to make use of mkimage
if needed.

Signed-off-by: Hans-Christian Egtvedt <hans-christian.egtvedt@atmel.com>