ext2: add possibility to build ext3 filesystem
This patch adds the possibility to select to build an ext3 filesystem. This is
done by adding a journal with the tune2fs tool. It is mandatory (so far) that
the build machine has this tool installed. The step will be skipped if tune2fs
is not installed and provide an error message to the user.

Signed-off-by: Hans-Christian Egtvedt <hans-christian.egtvedt@atmel.com>