AVR32: add JFFS2 image generation for ATSTK1006 board
The ATSTK1006 board has onboard 256 MB NAND flash, add to defconfig to generate
JFFS2 images to put into the NAND flash.

The flash images are divided into root and boot, where boot holds the kernel
and is supposed to be programmed into the NOR flash.

That way the kernel can be loaded in u-boot and then the kernel can initialize
the NAND flash and continue booting from the root file system.

Signed-off-by: Hans-Christian Egtvedt <hans-christian.egtvedt@atmel.com>