AVR32: add new addon board EVKLCD10X for ATNGW100
This patch adds support for EVKLCD10X (EVKLCD100 and EVKLCD101), which is an
addon board for ATNGW100.

This board has a QVGA (EVKLCD100) or VGA (EVKLCD101) touch screen LCD, and the
default configuration is targeted to show of this feature.

By default the configuration will build Qt Embedded with touch screen support.

Signed-off-by: Hans-Christian Egtvedt <hans-christian.egtvedt@atmel.com>