Add support for wget fallback to a mirror site
This patch adds a script which will let wget fallback to a mirror site if
failing to download source code.

The user will need to use scripts/wget.sh instead of executing wget directly,
and also supply a BR2_WGET_MIRROR in kconfig to the mirror site holding all the
source code tarballs.

Signed-off-by: Hans-Christian Egtvedt <hans-christian.egtvedt@atmel.com>