samba: bump version to 3.2.5
This patch bumps Samba version to 3.2.5. In the process Samba now
depends on libiconv. Some of the patches had to be updated for the new
version as well.

Two new patches are introduced, one to double check if BSD functions are
available, and one to skip checking for glibc version. The use unsigned
enum patch could be removed since this can be supplied with a autoconf
variable.

Signed-off-by: Hans-Christian Egtvedt <hans-christian.egtvedt@atmel.com>