atmel: add default Samba configuration and user credentials for NGW100 and NGW100 mkII
This patch adds the default Samba configuration for the NGW100 and NGW100 mkII
boards. It will setup to share a netdisk driver (/media on the system) to the
LAN side. In addition an user credentials file (smbpasswd) is also added. Only
the root user is added, using the default root/roota combo.

Signed-off-by: Hans-Christian Egtvedt <hans-christian.egtvedt@atmel.com>