target: add /boot directory to users of the generic device table
This patch adds the creation of the /boot directory to all targets that use the
generic device_table.txt and mini_device_table.txt file. The /boot directory is
typically used to store the Linux kernel, and other boot related files, and is
nice to get created on all systems, even if the directory is not populated.

/boot is defined by the Filesystem Hierarchy Standard (FHS).

Signed-off-by: Hans-Christian Egtvedt <hans-christian.egtvedt@atmel.com>