make: add a build successful message to top level Makefile
This patch adds a build successful message to the top level Makefile which is
printed upon successful builds. It will point the user to where the final
generated filesystem images are located.

For new users this is convenient to know that Buildroot actually completed
successfully and where the output files are located. All the output from
Buildroot might confuse users not used to scrolling build logs.

Signed-off-by: Hans-Christian Egtvedt <hans-christian.egtvedt@atmel.com>